Add files via upload
This repository contains the source code for the "Classic Studio" website, i created this for a web assignment focusing on HTML and CSS basics, with a touch of JavaScript for interactivity. The website features sections such as Home, Portfolio, Services, About, and Contact, showcasing a fictional photography business. It serves as a practical example of building a static website, demonstrating the application of fundamental web technologies to create a visually appealing and functional site. The primary purpose of this project is educational, aimed at helping beginners understand the basics of web development.